Fixed in users (user name and password lenght check, if the user name already exists, that the new user does not have the same name as the administrator). Added and updating Czech language translate for new strings. Fixing loading header (cpu temperature, load ...) -> redirection to update_footer.Added new tab "sensors". Added support for sensors. In the future it will be possible to add sensors (pressure, temperature, humidity, contact...) and their assignment to various operations (for example, the temperature sensor closes a certain program).

+++ ospy/helpers.py
@@ -365,7 +365,7 @@ def get_cpu_usage():
     prev_time_doing_nothing = time_doing_nothing
 
     # Output latest perccentage
-    return round(cpu_percentage ,1)  # To one digits     
+    return float(round(cpu_percentage ,1))  # To one digits     
 
 
 def mkdir_p(path):
@@ -625,6 +625,7 @@ def template_globals():
     from ospy.webpages import pluginStn
     from ospy import i18n
     from ospy.users import users
+    from ospy.sensors import sensors
 
     result = {
         'str': str,
